step1 refactor build
