+++ src/main/kotlin/domain/Row.kt
@@ -8,4 +8,22 @@ class Row(
     fun isMine(position: Int): Boolean {
         return cells[position].isMine()
     }
+
+    companion object {
+        fun build(
+            width: Int,
+            currentHeight: Int,
+            minesPosition: MinesPosition,
+        ): Row {
+            val widthRange = (0 until width)
+            val cells = widthRange.map { x ->
+                when (minesPosition.contains(Position(x, currentHeight))) {
+                    true -> MineCell()
+                    false -> GeneralCell()
+                }
+            }
+
+            return Row(cells)
+        }
+    }
 }
