+++ src/test/kotlin/domain/BoardTest.kt
@@ -1,5 +1,7 @@
 package domain
 
+import org.junit.jupiter.api.Assertions.assertAll
+import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
@@ -44,5 +46,28 @@ internal class BoardTest {
         val actual = board.isMine(Position(0, 0))
 
         assertFalse(actual)
+    }
+
+    @Test
+    fun `높이가 3 넓이가 2 마인의 위치가 ( (0,1), (1,2) )일때 (0,1), (1,2) 위치에 마인을 가진 3 * 2 의 지뢰찾기 보드가 생성된다`() {
+        val board = Board.build(
+            Width(2),
+            Height(3),
+            MinesPosition(
+                listOf(Position(0, 1), Position(1, 2))
+            )
+        )
+
+        assertAll(
+            "보드 생성 테스트",
+            { assertEquals(3, board.height) },
+            { assertEquals(2, board.width) },
+            { assertFalse(board.isMine(Position(0, 0))) },
+            { assertTrue(board.isMine(Position(0, 1))) },
+            { assertFalse(board.isMine(Position(0, 2))) },
+            { assertFalse(board.isMine(Position(1, 0))) },
+            { assertFalse(board.isMine(Position(1, 1))) },
+            { assertTrue(board.isMine(Position(1, 2))) },
+        )
     }
 }
